Renamed AcceptEula to ViewEulaSubscreen
