--- login-workflow/src/screens/subScreens/AcceptEula/index.ts
@@ -1,1 +1,0 @@
-export * from './AcceptEula';
